webpack 4 + Node 16からWXT + Vite + Vitest + Node 20に移行
- ビルドツール: webpack 4 → WXT (Vite 7ベース)
- テスト: Jest 29 → Vitest 4
- Node: 16 → 20 LTS
- CSS: Tailwind v2 (4.36MB) → v3 JIT (15.9KB)
- モジュール: CommonJS → ESM
- エントリポイント: ファイルベース規約 (src/entrypoints/)
