+++ mise.toml
@@ -1,2 +1,2 @@
 [tools]
-node = "16.20.2"
+node = "20.19.4"
