+++ package.json
@@ -3,64 +3,24 @@
   "version": "1.3.1",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack --config=webpack.dev.js --mode=development",
-    "build": "webpack --config webpack.prod.js",
-    "test": "jest"
+    "dev": "wxt",
+    "build": "wxt build",
+    "zip": "wxt zip",
+    "test": "vitest run"
   },
   "dependencies": {
-    "core-js": "^3.8.3",
-    "csv-loader": "^3.0.3",
-    "socket.io": "^3.1.2",
     "vue": "^3.2.37"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.3",
-    "@babel/preset-env": "^7.14.4",
-    "@tailwindcss/postcss7-compat": "^2.0.3",
-    "@vue/cli-plugin-babel": "^4.5.11",
-    "@vue/cli-plugin-eslint": "^4.5.11",
-    "@vue/cli-service": "^4.5.11",
-    "@vue/compiler-sfc": "^3.0.0",
-    "autoprefixer": "^9.8.6",
-    "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.2.2",
-    "clean-webpack-plugin": "^3.0.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.5.0",
-    "jest": "^29.7.0",
-    "postcss": "^7.0.35",
+    "@wxt-dev/module-vue": "^1.0.3",
+    "autoprefixer": "^10.4.27",
+    "postcss": "^8.5.8",
     "postcss-prefix-selector": "^1.7.2",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.3",
-    "vue-loader": "^16.1.2",
-    "webpack-cli": "^4.5.0",
-    "zip-webpack-plugin": "^4.0.1"
+    "tailwindcss": "^3.4.19",
+    "vitest": "^4.0.18",
+    "wxt": "^0.20.18"
   },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true,
-      "browser": true,
-      "webextensions": true
-    },
-    "extends": [
-      "plugin:vue/vue3-essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browser": {
-    "fs": false
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ],
   "volta": {
-    "node": "16.20.2"
+    "node": "20.19.4"
   }
 }
