+++ public/_locales/en/detailed-description.txt
