+++ public/_locales/ja/detailed-description.txt
