+++ public/_locales/ja/messages.json
