+++ public/_locales/zh_CN/detailed-description.txt
