+++ public/_locales/zh_TW/detailed-description.txt
