--- public/popup.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="ja">
-
-<head>
-    <meta charset="utf-8">
-</head>
-
-<body>
-  <div id="amzSchRoot"></div>
-</body>
-
-</html>
-
-<style>
-  body {
-  max-width: 600px;
-  min-width: 600px;
-  max-height: 600px;
-  min-height: 600px;
-}
-
-</style>
