+++ src/entrypoints/options/index.html
@@ -8,6 +8,7 @@
 
 <body>
   <div id="amzSchRoot"></div>
+  <script type="module" src="./main.js"></script>
 </body>
 
 </html>
