+++ src/entrypoints/popup/index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="ja">
+
+<head>
+    <meta charset="utf-8">
+    <style>
+      body {
+        max-width: 600px;
+        min-width: 600px;
+        max-height: 600px;
+        min-height: 600px;
+      }
+    </style>
+</head>
+
+<body>
+  <div id="amzSchRoot"></div>
+  <script type="module" src="./main.js"></script>
+</body>
+
+</html>
