+++ src/entrypoints/popup/main.js
@@ -1,0 +1,5 @@
+import { createApp } from "vue";
+import App from "@/popup/App.vue";
+import "@/index.css";
+
+createApp(App).mount("#amzSchRoot");
