--- src/manifest.json
@@ -1,65 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "__MSG_appName__",
-  "version": "1.3.1",
-  "description": "__MSG_appDesc__",
-  "author": "kbkn",
-  "icons": {
-    "48": "imgs/extension_icon48.png",
-    "128": "imgs/extension_icon128.png"
-  },
-  "default_locale": "en",
-  "content_scripts": [
-    {
-      "matches": ["https://game.mahjongsoul.com/*",
-      "https://mahjongsoul.game.yo-star.com/*",
-      "https://game.maj-soul.net/*",
-      "https://game.maj-soul.com/*"],
-      "js": ["content_script.js"]
-    },
-    {
-      "matches": ["https://game.mahjongsoul.com/*",
-      "https://mahjongsoul.game.yo-star.com/*",
-      "https://game.maj-soul.net/*",
-      "https://game.maj-soul.com/*"],
-      "js": ["main-world.js"],
-      "world": "MAIN"
-    },
-    {
-      "matches": ["https://naga.dmv.nico/naga_report/order_form/"],
-      "js": ["content_script_naga.js"]
-    },
-    {
-      "matches": ["https://mjai.ekyu.moe/*"],
-      "js": ["content_script_mjai.js"]
-    }
-  ],
-  "action": {
-    "default_popup": "popup.html"
-  },
-  "background": { "service_worker": "background.js" },
-  "permissions": ["storage"],
-  "host_permissions": [
-    "https://game.mahjongsoul.com/*",
-    "https://mahjongsoul.game.yo-star.com/*",
-    "https://game.maj-soul.net/*",
-    "https://game.maj-soul.com/*",
-    "https://naga.dmv.nico/naga_report/order_form/",
-    "https://mjai.ekyu.moe/"
-  ],
-  "options_page": "options.html",
-  "web_accessible_resources": [
-    {
-      "resources": [
-          "main-world.js"
-      ],
-      "matches": [
-          "https://game.mahjongsoul.com/*",
-          "https://mahjongsoul.game.yo-star.com/*",
-          "https://game.maj-soul.net/*",
-          "https://game.maj-soul.com/*"
-      ],
-      "use_dynamic_url": true
-    }
-  ]
-}
